moving all files to more logical folders to help get code base organized
