+++ src/core/entities/User.ts
@@ -1,4 +1,4 @@
-import { IUser } from "@/lib/interfaces/ISession";
+import { IUser } from "@/core/interfaces/ISession";
 
 export default class User implements IUser {
   name: string = "";
