+++ src/core/interfaces/IAbpAjaxResult.ts
