+++ src/core/interfaces/IAppConsts.ts
