+++ src/core/interfaces/IAuthorization.ts
