+++ src/core/interfaces/ISession.ts
