+++ src/core/interfaces/IUserLoginResult.ts
