+++ src/lib/ajax.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosInstance, AxiosRequestConfig } from "axios";
+import axios, { AxiosInstance } from "axios";
 import AppConsts from "./appconsts";
 
 const ajax: AxiosInstance = axios.create({
@@ -8,7 +8,7 @@ const ajax: AxiosInstance = axios.create({
 
 // tslint:disable-next-line:typedef
 ajax.interceptors.request.use(function(config) {
-  if (!!window.abp.auth.getToken()) {
+  if (!!abp.auth.getToken()) {
     config.headers.common.Authorization = "Bearer " + window.abp.auth.getToken();
   }
   return config;
