+++ src/services/interfaces/IAccountService.ts
@@ -1,4 +1,4 @@
-import { IUserLoginResult } from "@/lib/interfaces/IUserLoginResult";
+import { IUserLoginResult } from "@/core/interfaces/IUserLoginResult";
 
 export default interface ISessionService {
   login(username: string, password: string): Promise<IUserLoginResult>;
